Players, Matches and EndedMatches migrated to ElephantSQL (PostgreSQL), added 2-database support, changed Models and Repositories of migrated elements, changed id element from Long to Integer on every model, fixed small design bugs on results

+++ src/main/java/cstv/Controllers/GuideController.java
@@ -34,7 +34,7 @@ public class GuideController {
 
     @GetMapping("/guides/{id}")
     public ModelAndView getGuidePage(ModelAndView modelAndView,
-                                     @PathVariable Long id) {
+                                     @PathVariable Integer id) {
         try {
             Guide guide = guideService.findGuideById(id);
             Set<String> guideText = new TreeSet<>();
