+++ src/main/java/cstv/Models/DatabaseSequence.java
@@ -11,5 +11,5 @@ public class DatabaseSequence {
     @Id
     private String id;
 
-    private long seq;
+    private Integer seq;
 }
